+++ typings/index.d.ts
@@ -8813,6 +8813,10 @@ declare module 'coc.nvim' {
      */
     readonly dialog: boolean
     /**
+     * Is true when terminal feature is supported
+     */
+    readonly terminal: boolean
+    /**
      * Is true when vim's textprop is supported.
      */
     readonly textprop: boolean
@@ -9953,6 +9957,10 @@ declare module 'coc.nvim' {
      * An optional flag to make the picker accept multiple selections, if true the result is an array of picks.
      */
     canPickMany?: boolean
+    /**
+     * @deprecated use placeHolder instead.
+     */
+    placeholder?: string
   }
 
   /**