+++ history.md
@@ -2,6 +2,12 @@
 
 Notable changes of coc.nvim:
 
+## 2025-06-02
+
+- Add function keys support to notification popups on vim9.
+- Use notification dialogs with actions (instead of menu picker) when
+  'enableMessageDialog' is enabled.
+
 ## 2025-05-21
 
 - Perform format on save after execute `editor.codeActionsOnSave`, the same as